Separate Parser into just parser and parsed object
